Delete node from Circular Linked List.java
Algorithm----------------------------------------------
Define a Node class which represents a node in the list. It has two properties data and next which will point to the next node.
Define another class for creating the circular linked list and it has two nodes: head and tail. It has a variable size and two methods: deleteMid() and display() .
deleteMid() will delete the node from the middle of the list:
It first checks whethe the head is null (empty list) then, it will return from the function as there is no node present in the list.
If the list is not empty, it will check whether the list has only one node.
If the list has only one node, it will set both head and tail to null.
If the list has more than one node then, it will calculate the size of the list. Divide the size by 2 and store it in the variable count.
Temp will point to head, and current will point to the previous node to temp.
Iterate the list till current will point middle node of the list.
Current will point to node next to temp, i.e., removes the node next to current.
a. display() will show all the nodes present in the list.

Define a new node 'current' that will point to the head.
Print current.data till current will points to head again.
Current will point to the next node in the list in each iteration.