Transpose of a matrix
I was here for my making my first PR request for hacktoberfest. The transpose of a matrix is a new matrix that is obtained by exchanging the rows and columns.In this program, the user is asked to enter the number of rows r and columns c. Their values should be less than 10 in this program.Then, the user is asked to enter the elements of the matrix (of order r*c).