SORT WORDS IN ALPHABETIC ORDER
I was here for making a PR request for HacktoberFest 2020.In this program, we store the string to be sorted in my_str. Using the split() method the string is converted into a list of words. The split() method splits the string at whitespaces.The list of words is then sorted using the sort() method, and all the words are displayed.