Check Whether a String is Palindrome or Not
I was here for making PR request for Hacktoberfest 2020.In this program, we have taken a string stored in my_str.Using the method casefold() we make it suitable for caseless comparisons. Basically, this method returns a lowercased version of the string.We reverse the string using the built-in function reversed(). Since this function returns a reversed object, we use the list() function to convert them into a list before comparing.